+++ Krowi_AchievementFilter/Data/ExportedCategories.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2021-09-02 19-19-14 ]] --
+-- [[ Exported at 2021-10-16 14-20-14 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --
 
@@ -411,6 +411,7 @@ function exportedCategories.Load(a)
     tmp[8] = cat:New((EJ_GetInstanceInfo(240))); -- Wailing Caverns
     AddC(tmp[2], tmp[8]);
     AddA(tmp[8], a[630]); -- Wailing Caverns
+    AddA(tmp[8], a[11765]); -- Pet Battle Challenge: Wailing Caverns
     tmp[10] = cat:New((EJ_GetInstanceInfo(227))); -- Blackfathom Deeps
     AddC(tmp[2], tmp[10]);
     AddA(tmp[10], a[632]); -- Blackfathom Deeps
@@ -420,6 +421,7 @@ function exportedCategories.Load(a)
     tmp[12] = cat:New((EJ_GetInstanceInfo(231))); -- Gnomeregan
     AddC(tmp[2], tmp[12]);
     AddA(tmp[12], a[634]); -- Gnomeregan
+    AddA(tmp[12], a[13269]); -- Pet Battle Challenge: Gnomeregan
     tmp[13] = cat:New((EJ_GetInstanceInfo(233))); -- Razorfen Kraul
     AddC(tmp[2], tmp[13]);
     AddA(tmp[13], a[635]); -- Razorfen Kraul
@@ -442,6 +444,7 @@ function exportedCategories.Load(a)
     AddC(tmp[2], tmp[21]);
     AddA(tmp[21], a[642]); -- Blackrock Depths
     AddA(tmp[21], a[3496]); -- A Brew-FAST Mount
+    AddA(tmp[21], a[14020]); -- Pet Battle Challenge: Blackrock Depths
     tmp[22] = cat:New((EJ_GetInstanceInfo(229))); -- Lower Blackrock Spire
     AddC(tmp[2], tmp[22]);
     AddA(tmp[22], a[643]); -- Lower Blackrock Spire
@@ -457,6 +460,8 @@ function exportedCategories.Load(a)
     AddC(tmp[2], tmp[25]);
     AddA(tmp[25], a[646]); -- Stratholme
     AddA(tmp[25], a[729]); -- Deathcharger's Reins
+    AddA(tmp[25], a[13627]); -- Pet Battle Challenge: Stratholme
+    AddA(tmp[25], a[13766]); -- Malowned
     tmp[3] = cat:New((GetCategoryInfo(15271))); -- Raids
     AddC(tmp[1], tmp[3]);
     AddA(tmp[3], a[1285]); -- Classic Raider
@@ -504,6 +509,7 @@ function exportedCategories.Load(a)
     tmp[796] = cat:New((EJ_GetInstanceInfo(754)) .. " (" .. (GetCategoryInfo(15234)) .. ")"); -- Naxxramas
     AddC(tmp[3], tmp[796]);
     AddA(tmp[796], a[11744]); -- "Drop Dead, Gorgeous"
+    AddA(tmp[796], a[425]); -- "Atiesh, Greatstaff of the Guardian"
     tmp[785] = cat:New((GetCategoryInfo(15117))); -- Pet Battles
     AddC(tmp[1], tmp[785]);
     AddA(tmp[785], a[6586]); -- Eastern Kingdoms Safari
@@ -527,7 +533,7 @@ function exportedCategories.Load(a)
     AddA(tmp[616], a[902]); -- Chief Exalted Officer
     tmp[794] = cat:New(C_Map.GetMapInfo(111).name); -- Shattrath City
     AddC(tmp[616], tmp[794]);
-    AddA(tmp[794], a[1165]); -- "My Sack is ""Gigantique"""
+    AddA(tmp[794], a[1165]); -- "My Storage is ""Gigantique"""
     tmp[924] = cat:New(C_Map.GetMapInfo(110).name); -- Silvermoon City
     AddC(tmp[616], tmp[924]);
     AddA(tmp[924], a[613]); -- Killed in Quel'Thalas
@@ -1685,7 +1691,7 @@ function exportedCategories.Load(a)
     AddA(tmp[140], a[5505]); -- Bullet Time
     AddA(tmp[140], a[631]); -- Shadowfang Keep
     AddA(tmp[140], a[5093]); -- Heroic: Shadowfang Keep
-    AddA(tmp[140], a[4627]); -- Big Love Rocket
+    AddA(tmp[140], a[4627]); -- X-45 Heartbreaker
     tmp[141] = cat:New((EJ_GetInstanceInfo(76))); -- Zul'Gurub
     AddC(tmp[130], tmp[141]);
     AddA(tmp[141], a[5743]); -- It's Not Easy Being Green
@@ -2320,7 +2326,7 @@ function exportedCategories.Load(a)
     AddA(tmp[899], a[8063]); -- Heroic: Primordius
     AddA(tmp[899], a[8064]); -- Heroic: Dark Animus
     AddA(tmp[899], a[8065]); -- Heroic: Iron Qon
-    AddA(tmp[899], a[8066]); -- Heroic: Twin Consorts
+    AddA(tmp[899], a[8066]); -- Heroic: Twin Empyreans
     AddA(tmp[899], a[8067]); -- Heroic: Lei Shen
     AddA(tmp[899], a[8068]); -- Heroic: Ra-den
     tmp[206] = cat:New((EJ_GetInstanceInfo(369))); -- Siege of Orgrimmar
@@ -2380,7 +2386,7 @@ function exportedCategories.Load(a)
     AddA(tmp[549], a[7908]); -- I Choose You
     AddA(tmp[549], a[7936]); -- Pandaren Spirit Tamer
     AddA(tmp[549], a[8080]); -- Fabled Pandaren Tamer
-    tmp[790] = cat:New((GetCategoryInfo(15282)) .. " (" .. (GetCategoryInfo(15234)) .. ")"); -- Brawler's Guild
+    tmp[790] = cat:New(addon.L["Brawler's Guild"] .. " (" .. (GetCategoryInfo(15234)) .. ")"); -- Brawler's Guild
     AddC(tmp[189], tmp[790]);
     AddA(tmp[790], a[7937]); -- You Are Not Your $#*@! Legplates (Season 1)
     AddA(tmp[790], a[7940]); -- The Second Rule of Brawler's Guild (Season 1)
@@ -2962,7 +2968,7 @@ function exportedCategories.Load(a)
     AddA(tmp[548], a[9463]); -- Draenic Pet Battler
     AddA(tmp[548], a[9724]); -- Taming Draenor
     AddA(tmp[548], a[10052]); -- Tiny Terrors in Tanaan
-    tmp[791] = cat:New((GetCategoryInfo(15282)) .. " (" .. (GetCategoryInfo(15234)) .. ")"); -- Brawler's Guild
+    tmp[791] = cat:New(addon.L["Brawler's Guild"] .. " (" .. (GetCategoryInfo(15234)) .. ")"); -- Brawler's Guild
     AddC(tmp[223], tmp[791]);
     AddA(tmp[791], a[9168]); -- You Are Not Your $#*@! Legplates (Season 2)
     AddA(tmp[791], a[9169]); -- The Second Rule of Brawler's Guild (Season 2)
@@ -3592,7 +3598,7 @@ function exportedCategories.Load(a)
     AddA(tmp[430], a[12098]); -- Mechanical Melee
     AddA(tmp[430], a[12099]); -- Unstoppable Undead
     AddA(tmp[430], a[12100]); -- Family Fighter
-    tmp[792] = cat:New((GetCategoryInfo(15282)) .. " (" .. (GetCategoryInfo(15234)) .. ")"); -- Brawler's Guild
+    tmp[792] = cat:New(addon.L["Brawler's Guild"] .. " (" .. (GetCategoryInfo(15234)) .. ")"); -- Brawler's Guild
     AddC(tmp[237], tmp[792]);
     AddA(tmp[792], a[11560]); -- You Are Not Your $#*@! Legplates (Season 3)
     AddA(tmp[792], a[11561]); -- You Are Not Your $#*@! Legplates (Season 3)
@@ -4227,7 +4233,7 @@ function exportedCategories.Load(a)
     AddA(tmp[343], a[13625]); -- Mighty Minions of Mechagon
     AddA(tmp[343], a[13626]); -- Nautical Nuisances of Nazjatar
     AddA(tmp[343], a[13695]); -- Team Aquashock
-    tmp[793] = cat:New((GetCategoryInfo(15282)) .. " (" .. (GetCategoryInfo(15234)) .. ")"); -- Brawler's Guild
+    tmp[793] = cat:New(addon.L["Brawler's Guild"] .. " (" .. (GetCategoryInfo(15234)) .. ")"); -- Brawler's Guild
     AddC(tmp[258], tmp[793]);
     AddA(tmp[793], a[11558]); -- The First Rule of Brawler's Guild
     AddA(tmp[793], a[13186]); -- You Are Not Your $#*@! Legplates
@@ -4854,6 +4860,7 @@ function exportedCategories.Load(a)
     AddA(tmp[419], a[14629]); -- Gaining Respect
     AddA(tmp[419], a[14630]); -- Becoming a Hero
     AddA(tmp[419], a[14631]); -- Champion of the Covenant
+    AddA(tmp[419], a[15241]); -- Renowned
     AddA(tmp[419], a[14632]); -- Conducting Anima
     AddA(tmp[419], a[14633]); -- Master Navigator
     AddA(tmp[419], a[14636]); -- Adventurer in Chief
@@ -4924,6 +4931,11 @@ function exportedCategories.Load(a)
     AddA(tmp[420], a[14764]); -- The Great Luckydo
     AddA(tmp[420], a[14833]); -- Fashion Abomination
     AddA(tmp[420], a[14763]); -- Crypt Couture
+    tmp[952] = cat:New(addon.L["Other"]); -- Other
+    AddC(tmp[883], tmp[952]);
+    tmp[953] = cat:New((select(2, addon.GetAchievementInfo(14222)))); -- Exile's Reach
+    AddC(tmp[952], tmp[953]);
+    AddA(tmp[953], a[14222]); -- Exile's Reach
     tmp[857] = cat:New(addon.L["Excluded"]); -- Excluded
     AddC(tmp[883], tmp[857]);
     tmp[857].HasFlexibleData = true;
@@ -5130,8 +5142,8 @@ function exportedCategories.Load(a)
     AddA(tmp[917], a[1295]); -- Crashin' & Thrashin'
     AddA(tmp[917], a[279]); -- Simply Abominable
     AddA(tmp[917], a[1687]); -- Let It Snow
-    AddA(tmp[917], a[1685]); -- Bros. Before Ho Ho Ho's
-    AddA(tmp[917], a[1686]); -- Bros. Before Ho Ho Ho's
+    AddA(tmp[917], a[1685]); -- Holiday Bromance
+    AddA(tmp[917], a[1686]); -- Holiday Bromance
     AddA(tmp[917], a[1688]); -- The Winter Veil Gourmet
     AddA(tmp[917], a[1689]); -- He Knows If You've Been Naughty
     AddA(tmp[917], a[1690]); -- A Frosty Shake
@@ -5140,8 +5152,26 @@ function exportedCategories.Load(a)
     AddA(tmp[917], a[5853]); -- A-Caroling We Will Go
     AddA(tmp[917], a[5854]); -- A-Caroling We Will Go
     AddA(tmp[917], a[8699]); -- The Danger Zone
+    AddA(tmp[917], a[15181]); -- Rock n' Roll
     AddA(tmp[917], a[10353]); -- Iron Armada
     AddA(tmp[917], a[1691]); -- Merrymaker
+    tmp[950] = cat:New(addon.L["Gifts"], true); -- Gifts
+    AddC(tmp[917], tmp[950]);
+    AddA(tmp[950], a[1705]); -- Clockwork Rocket Bot
+    AddA(tmp[950], a[1706]); -- Crashin' Thrashin' Racer
+    AddA(tmp[950], a[6059]); -- Red Rider Air Rifle
+    AddA(tmp[950], a[6060]); -- MiniZep Controller
+    AddA(tmp[950], a[6061]); -- Gaudy Winter Veil Sweater
+    AddA(tmp[950], a[7852]); -- They're Both Footballs?
+    AddA(tmp[950], a[8698]); -- Crashin' Thrashin' Flyer
+    AddA(tmp[950], a[9915]); -- Crashin' Thrashin' Shredder
+    AddA(tmp[950], a[10337]); -- Crashin' Thrashin' Flamer
+    AddA(tmp[950], a[10751]); -- Brr... It's Cold in Here
+    AddA(tmp[950], a[11913]); -- Don't Play With That In The House
+    AddA(tmp[950], a[12850]); -- At Least You're Being Careful
+    AddA(tmp[950], a[13927]); -- Crashin' Splashin'
+    AddA(tmp[950], a[14931]); -- Flying Festivities
+    AddA(tmp[950], a[15313]); -- Rockin' Rollin' Racer
     tmp[925] = cat:New((GetCategoryInfo(15101))); -- Darkmoon Faire
     AddC(tmp[884], tmp[925]);
     AddA(tmp[925], a[6019]); -- "Come One, Come All!"
@@ -5204,6 +5234,14 @@ function exportedCategories.Load(a)
     AddA(tmp[925], a[11919]); -- Taking this Show on the Road
     AddA(tmp[925], a[11920]); -- Perfect Performance
     AddA(tmp[925], a[11921]); -- Mosh Pit
+    tmp[951] = cat:New((GetCategoryInfo(15454))); -- Timewalking
+    AddC(tmp[884], tmp[951]);
+    AddA(tmp[951], a[13931]); -- "Memories of Fel, Frost and Fire"
+    tmp[954] = cat:New((GetCategoryInfo(15258))); -- Legion
+    AddC(tmp[951], tmp[954]);
+    AddA(tmp[954], a[15309]); -- A Towering Success
+    AddA(tmp[954], a[15310]); -- A Tour of Towers
+    AddA(tmp[954], a[15308]); -- Tower Overwhelming
     tmp[947] = cat:New((GetCategoryInfo(15258))); -- Legion
     AddC(tmp[884], tmp[947]);
     tmp[948] = cat:New(addon.L["Legion Assaults"]); -- Legion Assaults
@@ -5225,6 +5263,7 @@ function exportedCategories.Load(a)
     AddA(tmp[935], a[13403]); -- Frontline Slayer
     tmp[936] = cat:New(addon.L["Assault on Drustvar"]); -- Assault on Drustvar
     AddC(tmp[933], tmp[936]);
+    AddA(tmp[936], a[13435]); -- Doomsoul Surprise
     AddA(tmp[936], a[13317]); -- Supplied and Ready
     AddA(tmp[936], a[13283]); -- Frontline Warrior
     AddA(tmp[936], a[13284]); -- Frontline Warrior