+++ Krowi_AchievementFilter/Data/ExportedNextPatchAchievements.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2021-08-10 07-55-07 ]] --
+-- [[ Exported at 2021-10-16 14-20-14 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --