+++ Krowi_AchievementFilter/GUI/SideButton.lua
@@ -6,6 +6,11 @@ local gui = addon.GUI;
 gui.SideButton = {};
 local sideButton = gui.SideButton;
 
+sideButton.OfsX1 = 0;
+sideButton.OfsY1 = 0;
+sideButton.OfsXn = 0;
+sideButton.OfsYn = 0;
+
 local OnClick;
 
 -- [[ Constructors ]] --
@@ -50,10 +55,10 @@ function sideButton:New(event, otherButtons)
                 if otherButtons[i].Event.ID == self.Event.ID then -- Found this button
                     if i == 1 and i + 1 <= #otherButtons then -- Button is the 1st and there are more
                         otherButtons[i + 1]:ClearAllPoints();
-                        otherButtons[i + 1]:SetPoint("TOPLEFT", AchievementFrame, "TOPRIGHT", 0, 0); -- Make the 2nd button anchor like the 1st one
+                        otherButtons[i + 1]:SetPoint("TOPLEFT", AchievementFrame, "TOPRIGHT", sideButton.OfsX1, sideButton.OfsY1); -- Make the 2nd button anchor like the 1st one
                     elseif i < #otherButtons then -- Button is somewhere in the middle
                         otherButtons[i + 1]:ClearAllPoints();
-                        otherButtons[i + 1]:SetPoint("TOPLEFT", otherButtons[i - 1], "BOTTOMLEFT", 0, 0); -- Make the 2nd button anchor like the 1st one
+                        otherButtons[i + 1]:SetPoint("TOPLEFT", otherButtons[i - 1], "BOTTOMLEFT", sideButton.OfsXn, sideButton.OfsYn); -- Make the 2nd button anchor like the 1st one
                     -- else -- Button is the last, nothing to move
                     end
                 end