+++ Krowi_AchievementFilter/CHANGELOG.md
@@ -1,6 +1,28 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [9.1.5.30.0] - 2021-11-02
+### Added
+- Winter Veil gift achievements under Events - Holidays - Winter Veil - Gifts
+- Legion Timewalking event and achievements under Events - Legion - Timewalking
+- Renowned achievement under Expansions - Shadowlands - Covenant Sanctums
+- Doomsoul Surprise achievement under Events - Battle for Azeroth - Faction Assaults - Assault on Drustvar
+- Pet Battle Challenge achievements to Expansions - Classic - Dungeons - Deadmines, - Wailing Caverns, - Gnomeregan, - Blackrock Depths and - Stratholme
+
+### Fixed
+- Brawler's Guild category was removed, added category name as localisable text
+- Unobtainable achievements now also have a red background color when ElvUI's Blizzard Skins and Achievements are selected
+- Event Reminders next to the Achievement window are now skinned when ElvUI's Blizzard Skins and Achievements are selected
+- Event Reminders pop ups are now skinned when ElvUI's Blizzard Skins and Alert Frames are selected
+
+[9.1.5.30.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3512817
+
+## [9.1.0.29.1] - 2021-09-07
+### Fixed
+- Map button overwrites Current Zone category and shows that one instead of Selected Zone 
+
+[9.1.0.29.1]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3453438
+
 ## [9.1.0.29.0] - 2021-09-07
 ### Added
 - Focus achievements